Create a repo card and refactored the code
