+++ cra-rxjs-styled-components/src/components/repo-card/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './RepoCard';
