--- cra-rxjs-styled-components/src/components/top-repositories/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './RepositoryCard';
