+++ cra-rxjs-styled-components/src/components/user-repos/UserRepos.tsx
@@ -1,19 +1,6 @@
-import { Star } from '@styled-icons/heroicons-outline';
-import {
-  Aside,
-  BadgeWrapper,
-  Containers,
-  Content,
-  Description,
-  Header,
-  HeadingLink,
-  RepoListWrapper,
-  StarBtn,
-  StarIcon,
-} from './UserRepos.styles';
-import RepoMeta from '../repo-meta';
-import PrivacyBadge from '../misc/Privacy-badge';
+import { RepoListWrapper } from './UserRepos.styles';
 import { useUserRepositories } from '../../hooks/user-repositories/use-user-repositories';
+import RepoCard from '../repo-card';
 
 function UserRepos({ isOrg = false }) {
   const { loading, repos } = useUserRepositories(isOrg);
@@ -24,48 +11,7 @@ function UserRepos({ isOrg = false }) {
 
   return (
     <RepoListWrapper>
-      {repos &&
-        repos.map(
-          ({
-            id,
-            name,
-            owner,
-            description,
-            stargazers_count,
-            forks_count,
-            language,
-            updated_at,
-            visibility,
-          }) => (
-            <Containers key={id}>
-              <Content>
-                <Header>
-                  <HeadingLink to={`/${owner.login}/${name}`}>
-                    {name}
-                  </HeadingLink>
-                  <BadgeWrapper>
-                    <PrivacyBadge visibility={visibility} />
-                  </BadgeWrapper>
-                </Header>
-                <Description>{description}</Description>
-                <RepoMeta
-                  language={language}
-                  forkCount={forks_count}
-                  stargazerCount={stargazers_count}
-                  updatedAt={updated_at}
-                />
-              </Content>
-              <Aside>
-                <StarBtn>
-                  <StarIcon>
-                    <Star />
-                  </StarIcon>
-                  Star
-                </StarBtn>
-              </Aside>
-            </Containers>
-          )
-        )}
+      {repos && repos.map((repo) => <RepoCard repo={repo} star />)}
     </RepoListWrapper>
   );
 }
