--- cra-rxjs-styled-components/src/hooks/user-repositories/types.ts
@@ -1,16 +1,0 @@
-export interface Repo {
-  id: string;
-  name: string;
-  owner: {
-    login: string;
-  };
-  description?: string | null;
-  languageColor?: string | null;
-  language?: string | null;
-  visibility: string;
-  isArchived: boolean;
-  isFork: boolean;
-  stargazers_count: number;
-  forks_count: number;
-  updated_at: any;
-}
