+++ cra-rxjs-styled-components/src/hooks/user-repositories/use-user-repositories.ts
@@ -1,12 +1,12 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { tap } from 'rxjs';
-import { Repo } from './types';
 import { ORG_REPO_LIST, USER_REPO_LIST } from '../../constants/url.constants';
 import { fromFetchWithAuth } from '../auth/from-fetch-with-auth';
+import { Repository } from '../../interfaces/repositories.interfaces';
 
 export function useUserRepositories(isOrg = false) {
-  const [repos, setRepos] = useState<Repo[]>();
+  const [repos, setRepos] = useState<Repository[]>();
   const [loading, setLoading] = useState(true);
   const { username } = useParams();
