+++ cra-rxjs-styled-components/src/interfaces/repositories.interfaces.ts
@@ -21,5 +21,5 @@ export interface Repository {
 }
 
 export interface RepositoryWithBranchCount extends Repository {
-  branches_count: number;
+  branches_count?: number;
 }
